Split to ImageList component
